slabs: no real changes, moving stuff around
