+++ slabs/downloads.go
@@ -3,6 +3,7 @@ package slabs
 import (
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"sort"
 	"strings"
@@ -15,15 +16,17 @@ import (
 	"go.uber.org/zap"
 )
 
+var errNotEnoughShards = errors.New("not enough shards")
+
 type downloadCandidates struct {
 	hosts   []hosts.Host
 	indices map[types.PublicKey]int
 }
 
-func newDownloadCandidates(allHosts []hosts.Host, slab Slab) downloadCandidates {
+func newDownloadCandidates(goodHosts []hosts.Host, slab Slab) downloadCandidates {
 	// build host lookup
-	lookup := make(map[types.PublicKey]struct{}, len(allHosts))
-	for _, h := range allHosts {
+	lookup := make(map[types.PublicKey]struct{}, len(goodHosts))
+	for _, h := range goodHosts {
 		lookup[h.PublicKey] = struct{}{}
 	}
 
@@ -42,7 +45,7 @@ func newDownloadCandidates(allHosts []hosts.Host, slab Slab) downloadCandidates 
 
 	// build sorted list of hosts, cheapest first
 	hosts := make([]hosts.Host, 0, len(lookup))
-	for _, h := range allHosts {
+	for _, h := range goodHosts {
 		if _, ok := indices[h.PublicKey]; ok {
 			hosts = append(hosts, h)
 		}
@@ -68,11 +71,11 @@ func (dc *downloadCandidates) next() (hosts.Host, bool) {
 
 // downloadShards downloads at least the minimum number of shards required to
 // recover the slab.
-func (m *SlabManager) downloadShards(ctx context.Context, slab Slab, allHosts []hosts.Host, logger *zap.Logger) ([][]byte, error) {
+func (m *SlabManager) downloadShards(ctx context.Context, slab Slab, goodHosts []hosts.Host, logger *zap.Logger) ([][]byte, error) {
 	shards := make([][]byte, len(slab.Sectors))
 	var downloaded atomic.Uint32
 
-	candidates := newDownloadCandidates(allHosts, slab)
+	candidates := newDownloadCandidates(goodHosts, slab)
 	if uint(len(candidates.indices)) < slab.MinShards {
 		return nil, fmt.Errorf("%w: only %d sectors available, minimum required: %d", errNotEnoughShards, len(candidates.indices), slab.MinShards)
 	}
