+++ slabs/manager_test.go
@@ -2,12 +2,15 @@ package slabs
 
 import (
 	"context"
+	"errors"
+	"io"
 	"slices"
 	"sync"
 	"time"
 
 	proto "go.sia.tech/core/rhp/v4"
 	"go.sia.tech/core/types"
+	"go.sia.tech/coreutils/rhp/v4"
 	"go.sia.tech/indexd/hosts"
 )
 
@@ -176,4 +179,76 @@ func (mock *mockhostManager) WithScannedHost(ctx context.Context, hk types.Publi
 		return hosts.ErrBadHost
 	}
 	return fn(host)
+}
+
+type mockDialer struct {
+	clients map[types.PublicKey]*mockHostClient
+}
+
+func newMockDialer(hosts []hosts.Host) *mockDialer {
+	clients := make(map[types.PublicKey]*mockHostClient, len(hosts))
+	for _, host := range hosts {
+		clients[host.PublicKey] = &mockHostClient{
+			sectors:  make(map[types.Hash256][proto.SectorSize]byte),
+			settings: host.Settings,
+		}
+	}
+	return &mockDialer{clients: clients}
+}
+
+func (d *mockDialer) DialHost(ctx context.Context, hostKey types.PublicKey, addr string) (HostClient, error) {
+	if client, ok := d.clients[hostKey]; ok {
+		return client, nil
+	}
+	return nil, errors.New("failed to dial host")
+}
+
+type mockHostClient struct {
+	delay    time.Duration
+	sectors  map[types.Hash256][proto.SectorSize]byte
+	settings proto.HostSettings
+}
+
+func (c *mockHostClient) ReadSector(ctx context.Context, prices proto.HostPrices, token proto.AccountToken, w io.Writer, root types.Hash256, offset, length uint64) (rhp.RPCReadSectorResult, error) {
+	select {
+	case <-time.After(c.delay):
+	case <-ctx.Done():
+		return rhp.RPCReadSectorResult{}, ctx.Err()
+	}
+
+	sector, ok := c.sectors[root]
+	if !ok {
+		return rhp.RPCReadSectorResult{}, proto.ErrSectorNotFound
+	}
+	_, err := w.Write(sector[:])
+	if err != nil {
+		return rhp.RPCReadSectorResult{}, err
+	}
+	return rhp.RPCReadSectorResult{
+		Usage: c.settings.Prices.RPCReadSectorCost(proto.SectorSize),
+	}, nil
+}
+
+func (c *mockHostClient) WriteSector(ctx context.Context, prices proto.HostPrices, token proto.AccountToken, data io.Reader, length uint64) (rhp.RPCWriteSectorResult, error) {
+	select {
+	case <-time.After(c.delay):
+	case <-ctx.Done():
+		return rhp.RPCWriteSectorResult{}, ctx.Err()
+	}
+
+	var sector [proto.SectorSize]byte
+	_, err := io.ReadFull(data, sector[:])
+	if err != nil {
+		return rhp.RPCWriteSectorResult{}, err
+	}
+	root := proto.SectorRoot(&sector)
+	c.sectors[root] = sector
+	return rhp.RPCWriteSectorResult{
+		Root:  root,
+		Usage: c.settings.Prices.RPCWriteSectorCost(proto.SectorSize),
+	}, nil
+}
+
+func (c *mockHostClient) Settings(context.Context, types.PublicKey) (proto.HostSettings, error) {
+	return c.settings, nil
 }
